[HaRepacker] Adjust ImageRenderViewer
Mostly fixes #6

Uses a Path element which removes the weird crosshair distortion from previous
Adjust calculation for the display rect so it no longer cuts off pixels
Crosshairs will now expand the display rect if moved to the edges
No longer shifts the image somewhat to the middle which resulted in weird zoom usage

Removed text saying width/height but that's cause I was too lazy to fix

+++ HaRepacker/Converter/ImageWidthOrHeightToScreenDPIConverter.cs
@@ -18,7 +18,7 @@ namespace HaRepacker.Converter {
 	public class ImageWidthOrHeightToScreenDPIConverter : IValueConverter {
 		public object Convert(object value, Type targetType, object parameter,
 			CultureInfo culture) {
-			var widthOrHeight = (double) value;
+			var widthOrHeight = (int) value;
 			var realWidthOrHeightToDisplay = widthOrHeight * ScreenDPIUtil.GetScreenScaleFactor();
 
 			return realWidthOrHeightToDisplay;
@@ -26,7 +26,7 @@ namespace HaRepacker.Converter {
 
 		public object ConvertBack(object value, Type targetType, object parameter,
 			CultureInfo culture) {
-			var value_ = (double) value;
+			var value_ = (int) value;
 			var imageWidthOrHeight = value_ / ScreenDPIUtil.GetScreenScaleFactor();
 
 			return imageWidthOrHeight;
